+++ src/elements/Order.php
@@ -2039,9 +2039,29 @@ class Order extends Element
             if (!$customer->id) {
                 throw new InvalidCallException('Customer must have an ID');
             }
+            $previousCustomerId = $this->customerId;
 
             $this->_customer = $customer;
             $this->customerId = $customer->id;
+
+            // If the customer is changing then we should be resetting the association with the addresses on the cart
+            if (($this->shippingAddressId || $this->billingAddressId) && $this->customerId != $previousCustomerId) {
+                if ($this->shippingAddressId && $shippingAddress = $this->getShippingAddress()) {
+                    $shippingAddress->id = null;
+                    $this->setShippingAddress($shippingAddress);
+                }
+
+                if ($this->billingAddressId && $billingAddress = $this->getBillingAddress()) {
+                    $billingAddress->id = null;
+                    $this->setBillingAddress($billingAddress);
+                }
+
+                $this->estimatedBillingAddressId = null;
+                $this->_estimatedBillingAddress = null;
+
+                $this->estimatedShippingAddressId = null;
+                $this->_estimatedShippingAddress = null;
+            }
         } else {
             $this->_customer = null;
             $this->customerId = null;